+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2020-10-13
+
 ### Added
 - The `:log` argument to the `:trace` option.
   Add a `trace: :log` option to your module's `use` statement,
@@ -73,7 +75,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic decision tree navigation
 
 
-[Unreleased]: https://github.com/Cantido/liberator/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/Cantido/liberator/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/Cantido/liberator/releases/tag/v1.3.0
 [1.2.0]: https://github.com/Cantido/liberator/releases/tag/v1.2.0
 [1.1.0]: https://github.com/Cantido/liberator/releases/tag/v1.1.0
 [1.0.0]: https://github.com/Cantido/liberator/releases/tag/v1.0.0